Add files via upload
This is a linear regression model I worked on  using the python libraries pandas, matplotlib, seaborn, and numpy. the program shows the correlation between price and square footage and displays this as a scatter plot. Then runs a linear regression using the data to predict what the price of new homes in the area alone with the error values for the model.